ai_agent_service/app/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
google-genai==1.52.0 
pydantic==2.9.2
